- fixed some issues with m_swap's design.
The native byte order converters were defined as macros which hid some issues due to lack of type checks.
Additionally the ???Long variants taking 'long' variables were removed, because longs are not always 32 bits so this could be destructive.
As a result of this, removed several DWORDs from struct definitions in i_crash.cpp.

+++ src/m_png.cpp
@@ -264,7 +264,7 @@ bool M_AppendPNGText (FileWriter *file, const char *keyword, const char *text)
 		{
 			crc = AddCRC32 (crc, (uint8_t *)text, len);
 		}
-		crc = BigLong((unsigned int)crc);
+		crc = BigLong(crc);
 		return file->Write (&crc, 4) == 4;
 	}
 	return false;
